+++ src/main/scala/tscfg/generators/Generator.scala
@@ -14,6 +14,10 @@ abstract class Generator(genOpts: GenOpts) {
   protected val className: String = genOpts.className
   protected val hasPath: String = if (genOpts.j7) "hasPath" else "hasPathOrNull"
   protected var genResults = GenResult()
+
+  protected def dbg(s: String, in: Boolean = true): String =
+    ""
+    //if (in) s" /*$s*/ " else s" // $s"
 }
 
 /**